Insert Range request header in half of the attempts
According to the developer of Restic, his code could be inserting a Range header into the request only asking for the first 10MB of the file.

As the file that is being requested is only 5MB, this shouldn’t matter. But, I thought I’d add it just to rule our the Range request being related to this issue.